More changes
Added Path and PathList to contain voice text
and file information. In particular, it contains pre and post pause
information.
This allows more accurate and flexible pausing to occur in the player.
Eliminates a bunch of redundant checking of voice_file path info.

Added ability to generate .txt files for engines which cache their
results. These .txt files can be used to generate the speech files in
the background or offline.