Many more changes
Focus on xml parsing
* Upgraded xml parsing to Python 3.
  Now uses ElementTree
* xml pre-processing of all .xml for the current skin is cached
read-only to avoid reprocessing on every new window being read or
re-read.
  Cache is pruned of all uninteresting nodes
  Indexes built.
  Old methods preserved for now to aide in debugging.
Kept an lxml version that is easier to use, but requires build of
lxml, so will end up discarding.

Prototyping a new way of scraping .xml files for voicing. Using only for
TTS settings/config. The existing method is limited by the lack of
accessible information in the .xml window scripts. The prototype adds
some additional xml elements to try to overcome a number of these
limitations. It will also serve as an example and means of communicating
what sort of information is needed for accessibility. The going is slow,
but very educational to me.

Fixed a number of problems/regressions in existing .xml scraping.

Reworked settings to share settings when possible. For example, there is
only one volume shared between voice generators. Each generator (engine)
has its own validator that can convert the shared volume to its own
scale. It will not be perfect, but no worse than before. But it will
also be much simpler.

For the initial release, will support only google and espeak engines. A
number of the other engines will be easy to get working again (flite,
festival, etc.). However, the more complex ones will take time. In
particular, SAPI will require changes to the Kodi embedded Python to
support. Either that or an awful hack added requiring SAPI to run on
it's own Python instance. The root problem is that Kodi uses
non-standard names for some of the Python libraries. This prevents
Windows 'comtype' libraries for binding with it. I have prototyped both
creating a Kodi build with the corrections as well as prototyping the
two-python instances solution.

The main work remaining is to complete the configuration work. It is
critical that it be completly accessible, which is why I have written
custom code for it since the vanilla configuration accessibility will be
less than perfect, at least initially.