Numerous changes (again)
VoiceCache reworked to handle caching of wave as well as mpg3 files.
Acknowledge that audio file type depends on the capabilities of
engine/player/transcoder as well as settings. Cache (and other code) now
takes this into account.
Added ability to ship with small subset of wave audio files that can be
used with the Kodi built-in player. This allows for user to get basic
configuration information even when the OS has now TTS engine or player.

Added Transcoder to convert from mp3 to wave or wave to mp3. Currently
focuses on using LAME.

Updated Settings Configuration to utilize these features.

Starting to add new mechanism to voice on Windows using native TTS
engine, etc. The previous method depended upon SAPI which requires
special Python addons which are not installable on current Kodi-built
Python. New model will use PowerShell instead of SAPI. Not sure if this
will give us all we need, but I believe it will.