Windows required langcodes hack
Kodi Python is built with Kodi and uses non-standard library/version
names. This means that .dlls which bind to python will not be able to
use Kodi Python. This broke langcodes library which supplies human
readable names to various locales. The work around is to use Dicts that
map locals to strings. The data is based on running Kodi. It is ugly and
prone to mistakes or changes over time.

Need better solution.