Fix issues with ServiceID
Some additional problems discovered with change over from representing
settings as strings to using ServiceID.
* Now all ServiceIDs must have a property. This resolves an ambiguity
with refering to an engine as 'engine'.'google' (with no property) to be
'engine'.'google'.'id'. This makes consistent with all other naming.

* Reworked SettingsMap and validators to contain more information
about a property, including validator, type and persistance info. The
goal is to centralize this information and to remove this type of info
from SettingsLowLevel, and other scattered places.

* Reworked Settings declarations to take advantage of these
ServiceID/SettingsMap changes.

Added items for product readiness.
 * Code now copies a basic keymap if one does not already exist.
 * Added settings to configure startup behavior:
 - display configuration on startup
 - display help on startup, including read-me
 - enable hint-text on-startup
 - enable more verbose configuration

Created powershell script to provide support to use powershell as a
speech engine.