Improve language config and loading settings
Testing revealed several related problems:

On startup, configured language variant and voice were reset to default
values.

Some settings were being accessed prior to being loaded into the
internal settings cache.

The first problem is resolved (but the
lang/voice code needs simplification).

A solution for the second problem has been
implemented, but only for the settings directly affected. Will likely
use the same solution for
all settings. It fits well with the effort to
simplify the handling of settings. The fix is to add detection of
accessing a setting which has not been yet loaded from settings.xml and
to provide a simple means to load individual settings at any time.