Config, usability and some bug fixes
The biggest change was to load settings earlier so that voicing could
start in time to voice some configuration actions/prompts.
Kodi caches the settings, both persisted settings from settings.xml as
well as triansient or constant settings. Was loading as engines,
players, and other stuff was being discovered. Changed to load purely on
demand of the setting. Did not require many real changes, but there was
enough small ones.

Also, verified and corrected the settings in settings.xml and which
settings are marked as persist. This hit a number of files which define
the settings. Small changes.

Since fader labels can not be scraped for text, created private means to
voice Notifications. Now used during configuration to clue the user in
on what is happening.

Got code to reload keymap after modification working again.

Windows native voicing was broken in several ways: Did not voice when
using SFX player. Did not get properly configured in several scenarios.
Resolved these issues.

On Windows, code to calculate md5 for textviewer required explict
encoding to utf-8.